requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
statsmodels>=0.14.0
scikit-learn>=1.3.0
python-docx>=0.8.11
openpyxl>=3.1.0
scipy>=1.10.0
